fix some error

+++ app/Http/Controllers/Dr/Panel/Turn/Schedule/ScheduleSetting/ScheduleSettingController.php
@@ -744,7 +744,9 @@ class ScheduleSettingController
       $isHoliday = true;
 
       $specialDay = SpecialDailySchedule::where('date',$validated['date'])->first();
-      $specialDay->delete();
+      if ($specialDay) {
+        $specialDay->delete();
+      }
       
 
     }
